Wait on FireStore document delete
FireStore delete() is async and the code didn't wait, so it was possible
for us to return from the delete with documents still existing.

Further, there was no handling of conflicts, so contention during delete
would not be retried. That _may_ account for the contention failure I
saw on DR-666.