Add retry to insert queries in the BigQueryPdao
This should greatly reduce the rateLimit errors we've been seeing the `snapshotByQueryHappyPathTest` test where we are doing more than 5 updates on a single dataset in a 10 second window.

The default values are to retry 3 times (so try up to 4 times) with a 500ms pause in between tries.  These values can be overridden