DR-2599 Eagerly create a GCS bucket when creating a dataset (#1292)
* Eagerly create a GCS bucket when creating a dataset

Do this rather than lazily creating on the first ingest.  This mitigates a race conditions where a bunch of first time ingests are triggered at the same time.  This makes sure that by the time the dataset is created and ready to use, it has a bucket to accept files

Note: we already follow this pattern for storage accounts and containers for Azure-backed datasets

* Fix test to test correct new behavior

* Make sure to delete projects on failed dataset/snapshot creation flights