Possible implementation of an IamAction cache that caches multiple actions per user instead of a single action. This also has some non-cache related changes to use `Set` instead of `List` and to use `IamAction` instead of `String` in a couple places.
When working on this I saw that there's a single authorization call chain that bypasses the cache, starting at `VerifyAuthorizationStep`. I don't think there's a need for this but wasn't sure, so I left it in place.

Note that the action cache code in `IamService.isAuthorized` has no coverage in unit tests. It might be good to add some as part of this work.