[DR-2673] Expand job endpoint permissions (#1324)
* Expand job visibility

If a user has permission on the
SAM resource associated with a job,
they can have view access to its
status and details.

* Fix spotbugs

* Group job access fields

* Revert "Group job access fields"

This reverts commit c2cd0f058e2c7c9a6163acafbf62367987c63f6a.

* Update remaining dataset jobs

* Add support for snapshot jobs

This excludes the snapshot export job
which will require a new SAM role
that is more specific.

* Add support for profile jobs

This excludes creating a snapshot
for now since it requires additional
work to determine who has access.

* Fix unit tests

* Fix job permissions test

* Fix job permission test

When running locally, the subject
id is the same for all test users. As
a result, the custodian test user had
access to the job even though they did
not have access to the dataset. This
change overrides the default subject id
to test the indended scenario.