Fix flight filtering
This was coming from an incorrectly named parameter in an annotation so I just removed most of the annotations and added unit tests.

I tried to upgrade to the latest and greatest openapi version but the various code generators don't have support.  As such, I simply annotated the default value.

I also added some unit tests to make sure that Spring is doing the right thing