Introduce SnapshotUpdateDuosDatasetFlight and steps
The flight can be used for setting, updating, and unsetting a snapshot's link to a DUOS dataset.
I broke the flight down into atomic steps for more straightforward undoing in the event of a flight failure.
The flight response is the DUOS Firecloud groups unlinked from and/or linked to the snapshot.