Add new IamAction.SHARE_POLICY_READER
If a caller is triggering a link/unlink of a DUOS dataset to/from a snapshot, they must hold this action on the snapshot.
At present, it is conferred by snapshot stewardship only.
This required special handling to account for the presence of colons in the Sam action name, a forbidden character in enum names.
Added unit tests.