Revert removal of SnapshotsApiController.RETRIEVE_INCLUDE_DEFAULT_VALUE
We set the default value of retrieveSnapshot endpoint's includes value in our OpenAPI spec, but our version of swagger-codegen 2.4.27 does not pull it into the generated SnapshotsApi.
This is because we define includes via a reference to a separately defined schema.
We've tried to upgrade to a 3.x version of swagger-codegen in the hopes of fixing this, but this broke other aspects of our code generation.
For now, we need to maintain a default value in the controller.