Add new endpoint syncDuosDatasetAuthorizedUsers
This allows TDR Admins to force a sync of a TDR-managed Firecloud group associated with the supplied DUOS ID.
If the DUOS ID is not known to TDR, throws a 404.
If we successfully update the Firecloud group's members but fail to record the sync in the DB due to a conflict, throws a 409.  (Note that in this case we do not walk back the sync: in reality, we don't expect this endpoint to be used heavily, and will instead move to automate a batch sync of all TDR-managed Firecloud groups.)