[DR-2793] Add endpoint to sync DUOS Firecloud group members (#1376)
* Add IamService.overwriteGroupPolicyEmails

Add new IamRole enum value MEMBER - used for managing Firecloud managed group members.
Added unit tests.

* Include DUOS in TDR service check as non-critical dependency

We call Consent's status endpoint unauthenticated from DuosClient within DuosService with exception handling.
Minor refactoring in StatusService for easier testing.
StatusTest was not a unit test: it made live calls to external services, so I expanded mocking and was able to remove one of our config faults in the process.
Expanded unit tests throughout.

* Introduce GoogleCredentialsService for obtaining default creds and tokens

This makes it easier to use mocks when  unit test code with which previously relied on GoogleCredentials static calls.
Refactored IamService to call these new service methods.
Added GoogleCredentialsService unit tests.

* Adjust DuosDaoTest to avoid use of tdrServiceAccountEmail bean

Obtaining this bean involves interaction with GoogleCredentials.

* Add DuosClient.getApprovedUsers which fetches approved users for a DUOS dataset

- Within user fetch, leverage GoogleCredentialsService to run as TDR SA
- Break out DUOS exception handling into common method - used for both dataset and authorized user fetch
- Add DuosClient unit tests

* Removed unneeded @VisibleForTesting annotation in standalone DuosDataset record

This record definition was previously embedded within DuosClient

* Add DuosDao.updateFirecloudGroupLastSyncedDate and unit test

* Add new endpoint syncDuosDatasetAuthorizedUsers

This allows TDR Admins to force a sync of a TDR-managed Firecloud group associated with the supplied DUOS ID.
If the DUOS ID is not known to TDR, throws a 404.
If we successfully update the Firecloud group's members but fail to record the sync in the DB due to a conflict, throws a 409.  (Note that in this case we do not walk back the sync: in reality, we don't expect this endpoint to be used heavily, and will instead move to automate a batch sync of all TDR-managed Firecloud groups.)

* Add DuosClient logging for all external calls save for status check

We just log the HTTP method, the URL, and a short description of the authorization used (unauthed, user-authed, or SA-authed -- no PII included)

* Fix DuosDaoTest.testUpdateFirecloudGroupLastSyncedDate

Previously failed in automation but not locally due to inconsistent truncation of Instant when fetched from DB

* Add DUOS endpoints retrieveDuosFirecloudGroup(s)

Expanded DuosDao, DuosService, and unit tests.

* Check for sync_duos_user action on datarepo before enabling sync

Now available in Sam dev: https://github.com/broadinstitute/sam/pull/907

* Update descriptions on syncAuthorizedUsers endpoint to more accurately reflect behavior

* Ensure that GoogleCredentialsService does not leak state among callers

Instead, make callers pass in requested scopes.

* Register DUOS status check as DUOS rather than Consent

Consent is the name of the API we call, but flagging an issue with DUOS is more likely to point us in the right direction, i.e. towards the DUOS team and Slack channel of the same name.

* Update IamRole.MEMBER with description clarifying its usage in Firecloud groups

It's a generic enough role name that it could be confused with IamRole.USER, which we reference frequently.

* Remove IamRole embedded description in favor of Javadoc comment for MEMBER

You then see the Javadoc comment when hovering on the reference in IntelliJ.

* Removed unused logger and @Api annotation in DuosApiController

The @Api annotation is unnecessary despite it being one that we use for our controllers: the overridden autogenerated API methods are tagged appropriately.

* Mocking exceptions in tests is overkill, we don't need to mock their methods

* Remove unneeded DuosApiController getters for request, object mapper

We see these in all of our controllers but they are only used in the default implementations of the api methods, which we override.
We can easily add them back in if we prefer that to be an explicit choice.

* Improvements to DuosClient and unit tests

- Downgrade logging to debug
- Concise switch statement rather than returning within when converting DUOS exceptions

* Downgrade most DuosService helper methods to private

Ensure unit test coverage for public methods also tests the expected outcomes from these calls
This loosens the coupling between internal implementation details and unit tests

* Favor obvious dummy scope names in GoogleCredentialsServiceTest

* Remove unused instance variable in IamService

* PR feedback for DuosDao and tests