[DR-2833] Introduce new endpoints for retrieving snapshot, dataset summaries (#1398)
* Add IamService.verifyAuthorization helper to throw if caller holds no actions on the resource

This follows the permission check we perform when determining what resources to enumerate: if the user is an authorized user of any kind on the resource, include it in the enumeration.

* Correct docstrings in IamService: auth verification methods throw a different exception

* SnapshotService now supports verifying a snapshot's readability or listability

These checks still fall back to checking for RAS passport-derived accessibility if the snapshot is inaccessible via SAM.
In all existing calls, we mean to check the snapshot's readability.
Expanded unit testing.

* Introduce new snapshots API endpoint retrieveSnapshotSummary

The caller is permitted to retrieve a snapshot summary if they can list the snapshot, so this matches the permission checks performed in snapshot enumeration.

* Introduce new datasets API endpoint retrieveDatasetSummary

The caller is permitted to retrieve a dataset summary if they can list the dataset, so this matches the permission checks performed in dataset enumeration.

* Remove info logs from snapshot controller's summary retrieval implementation

* Downgrade existing snapshot controller info logs to debug

This cuts on cost and the logs in question aren't that helpful in usual scenarios.  We can explicitly enable debug logging to view them.