Add SnapshotService.getSnapshotIdsAndRoles
For an authenticated user, this method returns the snapshot IDs accessible for the source TDR environment mapped to the roles which confer access.
Local benchmarking indicates that overall response time is primarily dictated by external calls to Sam and ECM:
our DB call to fetch TDR snapshot IDs is lightweight, and we've optimized our filtration of accessible snapshots down to those available in the given TDR environment by loading the TDR snapshot IDs into a HashSet first.
Streaming over the accessible snapshots rather than the TDR snapshot IDs is more efficient in production, where the former will be a subset of the latter.
Added unit test.