DRS test set-up disables BigQuery dataset accessibility check
We were still seeing significant test flakiness even with the 3 minute wait, on top of long test runtimes.
We decided that IAM propagation is unpredictable enough as to warrant disabling this check in automated tests.
Downstream test code does not interact directly with BigQuery: an earlier commit in this PR obtains a snapshot's DRS ID via TDR's snapshot preview endpoint.