[DR-2933] AzureIntegrationTest teardown would exit prematurely
When an AzureIntegrationTest run had 1+ snapshots to delete on teardown, the teardown method would not complete since we tried to remove items from a snapshot ID list while iterating over it via forEach.
Concretely, what this meant is that we would never get to the billing profile delete step in the teardown, and subsequent test runs in the same test suite would fail fast with a 'cannot reuse existing application deployment integrationapp from profile <x> with a different profile <y>' error.

- Updated the test teardown method to no longer remove elements from the snapshot ID list or null out instance-level resource IDs: these instance variables will all be reset in the test setup method.
- Moved contents of helper method clearEnvironment to test teardown method, and removed clearEnvironment: we should be handling this teardown in one place, rather than calling clearEnvironment at the end of test runs while also trying to clean up resources in the teardown method (which runs on both expected and premature test exits).
- Log billing profiles, snapshot summaries, and dataset summaries created in DataRepoFixtures for easier debugging in case of test error: none of these objects are large, and we might end up in a situation where we're debugging a test failure on a test environment that has since been repurposed for another test run.