[DR-2931] SamIam API construction moved to new SamApiService
This was done with the aim of making unit testing easier.
SamIamTest no longer has to spy the SamIam instance: we instead mock SamApiService and its responses.
Added unit tests to SamApiService to verify that our Sam ApiClient read timeouts are being set as we expect.

While here, I realized that RawlsServiceTest was involving SamIam unnecessarily.
It should only be testing RawlsServiceTest functionality.
I pared down this test substantially and added in RawlsService.getWorkspaceLink helper method to make testing easier.
WorkspaceResponse was refactored to no longer set a workspace link when serializing to a WorkspacePolicyModel:
it makes more sense to set the workspace link once we have the WorkspacePolicyModel in RawlsService.resolvePolicyEmails.