[DR-2843] Introduce new /scheduled endpoint for syncing DUOS group members
This endpoint should only be callable by the environment-specific SA.
But we could instead expand the authorization check in the existing DUOS sync endpoint to allow the TDR SA to call it (would sidestep the need to modify our oidc proxy settings).
This general pattern may not be suitable for future scheduled tasks, though.