Make DUOS group sync a Spring @Scheduled task
- By default, this cron is disabled: set DUOS_SYNCUSERS_SCHEDULE env var to a valid cron expression to schedule
- Manage distributed lock with ShedLock:
  - Minimally hold the lock for 5 seconds (this helps protect against node clock differences for short-running tasks)
  - We expect this task to take a few seconds
  - As a fallback in case the locking node dies, maximally hold the lock for 5 minutes (much longer than expected execution time)
Caveat: the lock will be unconditionally removed if present at the end of its maximum lock time, even if the task is still executing.
In this case the possibility of this happening and the risk if it does are low enough as to not warrant adding further complexity.
ShedLock also offers a KeepAliveLockProvider but advises its use only if necessary: https://github.com/lukas-krecan/ShedLock#keepalivelockprovider