[DR-2959] Merge and replace ingest rows get new IDs
If the ingest records for merge or replace ingests specify their own row IDs, we now ignore them and generate new ones.
This fixes a bug where row IDs specified in merge or replace ingests could result in both existing and new rows being soft-deleted.
This is because we soft-delete from a join on the raw and soft delete tables based on the row ID.
We couldn't unilaterally change this behavior because there are some historical active use cases where append ingests expect to be able to specify their own row IDs.