Add new endpoint getDatasetTags
This endpoint allows the caller to retrieve tags and their number of occurrences across all accessible datasets.
The caller can optionally filter by a query string (looks for tags which have a case-insensitive match) and restrict by a limit (all matches are returned if unspecified).
Results are returned first in descending order of occurrence count, then alphabetically by tag to be deterministic.
This is powered by new method TagDaoInterface.getTags for ease of reuse between datasets and snapshots.
Expanded unit testing in DatasetDaoTest.