Add new endpoint getSnapshotTags
This endpoint allows the caller to retrieve tags and their number of occurrences across all accessible snapshots (accessibility derived from Sam and a RAS passport check).
Introduced TagCountResultModel to also return any errors encountered when constructing our tag list: this is a pattern we follow for snapshot retrieval operations to account for the possibility of partial results, such as if we can successfully query Sam but not ECM in our passport check.
To be consistent among resources I ported getDatasetEndpoint over to return a TagCountResultModel too.
We can rely on existing unit tests for SnapshotService.getAccessibleSnapshots and DatasetDao.getTags (which uses TagDaoInterface.getTags default implementation, as does SnapshotDao).