ValidateBucketAccessStep can validate bulk file array ingest
Such files are directly specified in the ingest request's loadArray.
We don't clear them from the ingest request like we do when setting up for a combined ingest, so we'll have access to them. (For combined ingests, TDR is first writing them to a scratch control file and later validating their access and format as if they were ingested with JSON format.)
Added comments throughout to disambiguate.