Change Dataset helper to hasDedicatedGcpServiceAccount
This will better handle the in-flight case where GCP files are ingested to an Azure dataset:
The source dataset would have Azure as its platform and it would use the general TDR SA to facilitate the ingest,
so we want to make sure that we verify that the caller can read the GCP files.
Normally, we don't need to verify a caller's permission on Azure files to ingest because we use signed URLs for those.
Also updated Swagger documentation to reflect new behavior and explain to our users why we're advocating for dedicated SAs.