ValidateBucketAccessStep should accept CloudFileReader instead of GcsPdao
GcsPdao implements the CloudFileReader interface so this change is backwards-compatible.
This allows us to pass either a GcsPdao or AzureBlobStorePdao to the step, determined by the platform of the ingest's destination dataset.
If a user is attempting to ingest GCS files to an Azure dataset, this step would them call AzureBlobStorePdao.validateUserCanRead which is already set up to handle that case.