[DR-2975] GCP datasets with dedicated SAs don't need to check file perms (#1453)
* Add dedicatedServiceAccount flag to GoogleProjectResource

Populate it from GoogleResourceDao: it's true if the resource has a non-empty SA that doesn't match the general SA.
(Right now, the behavior is that a resource's SA is empty in the DB if it uses the general SA.)
Expanded / refactored unit tests.

* Add Dataset.shouldValidateIngesterFileAccess helper method

This method uses GoogleProjectResource.dedicatedServiceAccount to determine whether GCP datasets necessitate validating ingester file permissions.
Added unit test.

* Add dataset parameter to CloudFileReader.validateUserCanRead

The GcsPdao implementation is now a no-op if the dataset does not necessitate file permission checking.
Wired the dataset in throughout.

* Drive-by: use pattern matching for instanceof

* GCP bulk file ingests, JSON format, bulk mode should verify file access

Achieved via IngestUtils refactor: new method validateBulkFileLoadModelsFromStream is called within IngestBulkGcpBulkFileStep.
Moved ValidateBucketAccessStep to run for all FileIngestBulkFlight runs with GCP platform:
previously this only ran for GCP bulk file ingests, JSON format, non-bulk mode.

We recognize that there is still a gap in access validation for all GCP array ingests across flights.
This will be ticketed separately.

* ValidateBucketAccessStep can validate bulk file array ingest

Such files are directly specified in the ingest request's loadArray.
We don't clear them from the ingest request like we do when setting up for a combined ingest, so we'll have access to them. (For combined ingests, TDR is first writing them to a scratch control file and later validating their access and format as if they were ingested with JSON format.)
Added comments throughout to disambiguate.

* Add unit tests for ValidateBucketAccessStep

* Change Dataset helper to hasDedicatedGcpServiceAccount

This will better handle the in-flight case where GCP files are ingested to an Azure dataset:
The source dataset would have Azure as its platform and it would use the general TDR SA to facilitate the ingest,
so we want to make sure that we verify that the caller can read the GCP files.
Normally, we don't need to verify a caller's permission on Azure files to ingest because we use signed URLs for those.
Also updated Swagger documentation to reflect new behavior and explain to our users why we're advocating for dedicated SAs.

* Correct grammar in test assertion

* Test fixture refactoring for dedicated SA datasets

This will make it easier to grant ingest buckets the necessary permissions in tests, as well as to perform clean-up specific to datasets with dedicated SAs (revoke ingest bucket permissions, delete SA from Terra).

* ValidateBucketAccessStep may run for GCP ingests to Azure datasets

In preparation for calling this step for Azure datasets, we only verify GCP source paths for readability.
Expanded step unit tests.

* FileIngestFlight could ingest a GCP file to an Azure dataset

So include ValidateBucketAccessStep no matter the platform of the destination dataset.
Added unit test (in JUnit 5) for flight construction: this ensures that we're including the necessary verification step for all expected platforms.

* FileIngestBulkFlight could ingest GCP files to an Azure dataset

So include ValidateBucketAccessStep no matter the platform of the destination dataset.
Added unit test (in JUnit 5) for flight construction: this ensures that we're including the necessary verification steps for all expected platforms, formats (array or JSON control file), and bulk mode designation.

* Forgot to add new FlightTestUtils

This contains a helper method for getting the step names of a flight.

* Self-hosted dataset integration tests also require legacy bucket reader role for dedicated SA

Otherwise, attempts to get DRS objects fail with '<dedicated-sa> does not have storage.buckets.get access to the Google Cloud Storage bucket. Permission 'storage.buckets.get' denied on resource' message.

* Remove unused import in FlightTestUtils

* ValidateBucketAccessStep should accept CloudFileReader instead of GcsPdao

GcsPdao implements the CloudFileReader interface so this change is backwards-compatible.
This allows us to pass either a GcsPdao or AzureBlobStorePdao to the step, determined by the platform of the ingest's destination dataset.
If a user is attempting to ingest GCS files to an Azure dataset, this step would them call AzureBlobStorePdao.validateUserCanRead which is already set up to handle that case.

* Update CloudFileReader passed to ValidateBucketAccessStep in FileIngestFlight, FileIngestBulkFlight

I had previously pulled the step out in these two flights to run unconditionally no matter the platform of the destination dataset.

* Drive-by: convert previously added unit test to JUnit 5

* Clarify documentation on ValidateBucketAccessStep

* Unconditionally run ValidateBucketAccessStep in DatasetIngestFlight

Previously, we only ran it for GCP JSON-type (JSON or ARRAY format) ingests.
This was an oversignt as it meant that we weren't validating accessibility of GCP CSV control files.
We don't necessarily need to run it for Azure ingests because we don't think we want users trying to specify GCP control file for Azure ingests, and this step validates accessibility of directly specified files (in this case, the control file for JSON or CSV format ingests)...
but there's no harm in running it unconditionally as it will amount to a no-op if we have an Azure destination dataset.
(Adding this step conditionally has lead to us missing a few cases where it should have been running.)

Drive-by: refactored IngestUtilsTest to use JUnit 5, add coverage for isJsonTypeIngest, and downgrade to a pure unit test without mocking or unneeded costly set-up.