Retrieve datasets with no regard for exclusive locks present
In any flight where we attempt a modification of the dataset, we try to obtain a lock on it: if an exclusive lock is present, the flight will fail and the modification will not be allowed to proceed.
This changeset allows datasets to be retrieved in read-only operations even when exclusively locked.
Removed now-unneeded code and updated tests.
Refactored DatasetsApiControllerTest to JUnit 5 with test slicing.

Drive-by: refactored out DatasetService method for checking that callers can read datasets, and instead call the IamService within DatasetsApiController.  This code followed a pattern set in SnapshotService but was overly complex for datasets (a user may be able to read a snapshot because they have direct access via Sam and/or indirect access via a RAS passport... dataset access is only handled by Sam).