Retrieve snapshots with no regard for exclusive locks present
In any flight where we attempt a modification of the snapshot, we should obtain a lock on it: if an exclusive lock is present, the flight will fail and the modification will not be allowed to proceed.
This changeset allows snapshots to be retrieved in read-only operations even when exclusively locked.
Removed now-unneeded code and updated tests.

Improvement: added new enum LockOperation to allow for more descriptive error messages to users when their jobs fail on trying to lock a resource (in an upcoming commit, I will refactor DatasetDao to use this).
Drive-bys: addressed Intellij warnings where they were a small lift.