Snapshot creation flight should obtain shared lock on source dataset
Awhile back, snapshot creation would obtain an exclusive lock on its source dataset:
this was overkill and caused problems for reasonable concurrent operations performed on the same dataset.
It was removed in full.
However, snapshot creation should obtain a shared (non-exclusive) lock on its source dataset to guard against the dataset being deleted out from under it, which would result in an orphaned snapshot (we don't allow dataset deletion to proceed if the dataset has snapshots).