PR feedback: unit tests to verify that snapshot deletion locks the snapshot
In a previous PR, I removed a flaky ignored connected test that ran two concurrent snapshot deletes.
Talking with @snf2ye raised the million-dollar question: 'What exactly where we trying to test with that flaky test?'

The answer: we wanted to make sure that snapshot deletion takes out an exclusive lock on the snapshot, and that concurrent attempts at snapshot deletion would fail as a result.
A more complete, precise, stable, and speedy way to test this is through a chain of unit tests:
- Verify that SnapshotDeleteFlight adds LockSnapshotStep to its manifest in the way we expect
- Verify that LockSnapshotStep behaves as we'd expect for all possible input conditions (including if the snapshot is already locked, has already been deleted, etc).
- Verify that the underlying DAO methods to lock / unlock snapshots behave as we'd expect

We already had unit tests to verify DAO behavior.  I supplemented tests to include unit tests for flight construction and step behavior.  For symmetry, I also wrote unit tests for UnlockSnapshotStep, even though it's not called in this flight.

Future similar opportunities in 'test engoodification' include similar tests for dataset locking / unlocking steps, as well as flight construction in other places to verify locking behavior.