[DR-268] Resources can be fetched when locked (#1470)
* Add lockingJobId to dataset, snapshot response objects

This field will be included on resource retrieval, summary retrieval, and enumeration.
At this point, we only intend on exposing flight IDs which have obtained exclusive locks on resources:
flight IDs which have obtained shared locks on resources are still internal-only.

* Add lockingJobId to dataset summary object

Expose it from a getter in the dataset object and use it for conversions to model objects used in API responses.

* Retrieve datasets with no regard for exclusive locks present

In any flight where we attempt a modification of the dataset, we try to obtain a lock on it: if an exclusive lock is present, the flight will fail and the modification will not be allowed to proceed.
This changeset allows datasets to be retrieved in read-only operations even when exclusively locked.
Removed now-unneeded code and updated tests.
Refactored DatasetsApiControllerTest to JUnit 5 with test slicing.

Drive-by: refactored out DatasetService method for checking that callers can read datasets, and instead call the IamService within DatasetsApiController.  This code followed a pattern set in SnapshotService but was overly complex for datasets (a user may be able to read a snapshot because they have direct access via Sam and/or indirect access via a RAS passport... dataset access is only handled by Sam).

* Spotless and spotbugs

* Add lockingJobId to snapshot, snapshot summary objects

* Retrieve snapshots with no regard for exclusive locks present

In any flight where we attempt a modification of the snapshot, we should obtain a lock on it: if an exclusive lock is present, the flight will fail and the modification will not be allowed to proceed.
This changeset allows snapshots to be retrieved in read-only operations even when exclusively locked.
Removed now-unneeded code and updated tests.

Improvement: added new enum LockOperation to allow for more descriptive error messages to users when their jobs fail on trying to lock a resource (in an upcoming commit, I will refactor DatasetDao to use this).
Drive-bys: addressed Intellij warnings where they were a small lift.

* Refactor DatasetDao to used shared LockOperation enum

This allows for more descriptive error messages to users when their jobs fail on trying to lock a dataset.
Drive-bys: addressed Intellij warnings where they were a small lift (ex. removing unused userReq param).

* Resources written as job responses should be done after unlock

Otherwise, we return a resource summary object that could incorrectly indicate that it's exclusively locked.
This was exposed by broken connected tests.

* SnapshotService.makeSourceModelFromSource should call DatasetSummary.toModel

The previous commit exposed a latent bug: inconsistency in how a dataset source model was constructed, which caused several connected tests to fail due to mismatched values.
In the process of investigating these test failures, I found that SnapshotFileLookupConnectedTest could have several tests removed as long as we increased SnapshotDao unit test coverage.
I made the same test expansion for DatasetDao unit tests.
A nice side effect of this test refactor is that we got to remove two configuration faults previously used for testing only.

* Snapshot creation flight should obtain shared lock on source dataset

Awhile back, snapshot creation would obtain an exclusive lock on its source dataset:
this was overkill and caused problems for reasonable concurrent operations performed on the same dataset.
It was removed in full.
However, snapshot creation should obtain a shared (non-exclusive) lock on its source dataset to guard against the dataset being deleted out from under it, which would result in an orphaned snapshot (we don't allow dataset deletion to proceed if the dataset has snapshots).

* Spotbugs

* PR feedback: unit tests to verify that snapshot deletion locks the snapshot

In a previous PR, I removed a flaky ignored connected test that ran two concurrent snapshot deletes.
Talking with @snf2ye raised the million-dollar question: 'What exactly where we trying to test with that flaky test?'

The answer: we wanted to make sure that snapshot deletion takes out an exclusive lock on the snapshot, and that concurrent attempts at snapshot deletion would fail as a result.
A more complete, precise, stable, and speedy way to test this is through a chain of unit tests:
- Verify that SnapshotDeleteFlight adds LockSnapshotStep to its manifest in the way we expect
- Verify that LockSnapshotStep behaves as we'd expect for all possible input conditions (including if the snapshot is already locked, has already been deleted, etc).
- Verify that the underlying DAO methods to lock / unlock snapshots behave as we'd expect

We already had unit tests to verify DAO behavior.  I supplemented tests to include unit tests for flight construction and step behavior.  For symmetry, I also wrote unit tests for UnlockSnapshotStep, even though it's not called in this flight.

Future similar opportunities in 'test engoodification' include similar tests for dataset locking / unlocking steps, as well as flight construction in other places to verify locking behavior.

* Add unit tests to verify resource locking in snapshot creation flight

* Return dataset exclusive lock in encapsulating ResourceLocks structure

* Return snapshot exclusive lock in encapsulating ResourceLocks structure

* Remove duplicate test set-up in DatasetConnectedTest.testProjectDeleteAfterDatasetDelete

This set-up is already performed before the test and triggered a spotbugs error

* Remove DatasetConnectedTest.testOverlappingDeletes in favor of unit tests

Added DatasetDeleteFlightTest -- we already had unit test coverage for step behavior, DAO behavior.
When moving to remove the faults referenced in that connected test, I found that they were used in DatasetLockConnectedTest too.
But that test suite was not annotated, so its tests hadn't been running for 2 years.
I went through those tests and expanded flight construction unit tests to make up the coverage.
I was then able to remove 9 faults, clean up everywhere they were referenced in production code, and remove DatasetLockConnectedTest suite in full.

* @pshapiro4broad PR feedback

Including:
- Enum fields follow Java static final naming convention
- Consolidate calls to DatasetDao: some places we previously fetched a dataset to then convert it to a model, but we can get this in a single call now
- Use Optional.orElseThrow to concisely return optional content, or throw if not present