[DR-2992] Pin json-smart to 2.4.10 to address CVE-2023-1370 (#1486)
The affected version is 2.4.8 and the maintainers recommend upgrading to 2.4.10 for an adgacent bug fix.
json-smart is a transitive dependency from azure-identity and Spring Boot.
We recently upgraded azure-identity which gave us the necessary json-smart bump, but a larger Spring Boot upgrade that would have given us the json-smart bump for free was not working out of the box.
For now, we'll pin json-smart to 2.4.10 and note that it can be removed when we next upgrade Spring Boot.

As a drive-by, we don't need to pin log4j2 to mitigate the Log4Shell vulnerability anymore.  We are naturally getting Log4J v2.17.2 with our current Spring Boot.

I verified the json-smart and log4j versioning with Gradle dependencyInsight task.