[DR-3175] Make sure that Azure datasets properly truncate int fields on ingest (#1508)
* Make sure that Azure datasets properly truncate int fields on ingest

When ingesting from a CSV file into an Azure dataset, if you have an int field but the data has a period, the ingest fails.  That's too brittle.  The behavior should be to truncate the field (that's the behavior on GCP).  This PR fixes that by treating int and int64 field values as numeric (instead of int and bigint)

I modified an integration test to validate (int made sense to me since this is testing behavior of a third party system)

* Ensure that the signed URL for the tabular data scratch file is long enough

This PR changes the signed URL form 1H to the default (24h).  This was causing issues on bigger ingests

* Use cleaner switch syntax

* PR feedback and set the numeric size for ints to working values