[DR-3315] Snapshot auth domain can't be set when snapshot already has a policy
1. When creating a PAO, if TPS throws a 400 error with messaging indicating a duplicate PAO exception, fall back to updating the existing PAO.
We previously only fell back when encountering a 409 error from TPS, which their Swagger documentation indicates is the expected response when trying to create a PAO for an object that already exists.
In reality, TPS throws a 400 error: ID-899 was filed to reclassify the error correctly, but in the mean time we've made sure that TDR handles the current case.

2. When updating a PAO, TDR must specify the update mode.
Their Swagger documentation indicates that this is a required field.
Any time TDR updates a PAO, we now set the update mode to FAIL_ON_CONFLICT.

Updated unit tests accordingly, and added comments indicating opportunities for future cleanup.