Simplify Provider states to the minimum required
This reduces ongoing maintenance burden and better reflects the expectations that WDS has of TDR.
Upgraded to latest compatible pact plugins.
Moved away from deprecated host and scheme system properties in favor of setting pactbroker.url.
Since this is getting set in a test profile, we should no longer need to pass it as an environmentn variable in GHA verify_consumer_pacts.