Add consumerVersionSelectors to Pact provider set-up
This makes it so that Pact provider verification will pick up on the expected consumer Pacts from the broker.
Included documentation on how to expand the selector if you wish to pick up Pacts from a consumer's feature branch.
In coordination with WDS consumer, expanded 'user has access to snapshot with given id' provider state to include a table:
this consumer expects that a snapshot retrieved with tables will have at least one table with a name.