Add DUOS-ID filter option when enumerating snapshots
Add the ability to filter on a list of DUOS IDs.  Also return the duos id of a snapshot as part of the snapshot summary.

This is to enable functionality in the DUOS UI that will allow seeing which snapshots are linked to to which DUOS datasets and will allow exporting to Terra from DUOS.