[DR-3399] Fix NPE when DRS PostObject endpoint called without expand
OpenAPI generates a DRSPassportRequestModel for us with a Boolean (capital b :-) field called 'expand'.
If our DRS PostObject endpoint is called without 'expand' set in this response body (as is the case with calls from DrsHub), we'd get an NPE when trying to convert the Boolean to a boolean.
Added null handling to default to false for consistency with our DRS GetObject endpoint, and updated unit test set-up to reflect this.
(Before the fix, the unit test set-up change rightfully showed NPEs and failing tests.)